+++ api/src/routes.ts
@@ -1,5 +1,13 @@
 import express from 'express';
-import { createUsername, getUsernames, refreshScore, getUsernameById, searchUsernames } from '@/controller';
+import {
+  createUsername,
+  getUsernames,
+  refreshScore,
+  getUsernameById,
+  searchUsernames,
+  getLatestChatMessages,
+  getOlderChatMessages,
+} from '@/controller';
 
 export const router = express.Router();
 
@@ -8,3 +16,6 @@ router.get('/', getUsernames);
 router.get('/refresh/:id', refreshScore);
 router.get('/single/:id', getUsernameById);
 router.get('/search', searchUsernames);
+
+router.get('/chat/messages', getLatestChatMessages);
+router.get('/chat/messages/:oldestMessageId', getOlderChatMessages);